<fix> update data_loader -> for serving
