+++ dkt/dataloader.py
@@ -7,13 +7,13 @@ import random
 from sklearn.preprocessing import LabelEncoder
 import numpy as np
 import torch
+from torch.nn.utils.rnn import pad_sequence
 
 class Preprocess:
     def __init__(self,args):
         self.args = args
         self.train_data = None
         self.test_data = None
-        
 
     def get_train_data(self):
         return self.train_data
@@ -39,7 +39,7 @@ class Preprocess:
         le_path = os.path.join(self.args.asset_dir, name + '_classes.npy')
         np.save(le_path, encoder.classes_)
 
-    def __preprocessing(self, df, is_train = True):
+    def __preprocessing(self, df, is_train=False):
         cate_cols = ['assessmentItemID', 'testId', 'KnowledgeTag']
 
         if not os.path.exists(self.args.asset_dir):
@@ -47,36 +47,56 @@ class Preprocess:
             
         for col in cate_cols:
             
-            
             le = LabelEncoder()
+            
             if is_train:
-                #For UNKNOWN class
+                # for unknown class
                 a = df[col].unique().tolist() + ['unknown']
                 le.fit(a)
-                self.__save_labels(le, col)
+                self.__save__labels(le,col)
             else:
                 label_path = os.path.join(self.args.asset_dir,col+'_classes.npy')
                 le.classes_ = np.load(label_path)
-                
-                df[col] = df[col].apply(lambda x: x if x in le.classes_ else 'unknown')
+                df[col] = df[col].apply(lambda x: x if str(x) in le.classes_ else 'unknown')
 
             #모든 컬럼이 범주형이라고 가정
             df[col]= df[col].astype(str)
             test = le.transform(df[col])
             df[col] = test
-            
 
         def convert_time(s):
             timestamp = time.mktime(datetime.strptime(s, '%Y-%m-%d %H:%M:%S').timetuple())
             return int(timestamp)
 
         df['Timestamp'] = df['Timestamp'].apply(convert_time)
-        
+        print('Preprocess Done....')
         return df
 
     def __feature_engineering(self, df):
         #TODO
         return df
+
+    def load_data(self, df):
+        
+        df = self.__feature_engineering(df)
+        df = self.__preprocessing(df)
+        
+        # 추후 feature를 embedding할 시에 embedding_layer의 input 크기를 결정할때 사용
+        self.args.n_questions = len(np.load(os.path.join(self.args.asset_dir,'assessmentItemID_classes.npy')))
+        self.args.n_test = len(np.load(os.path.join(self.args.asset_dir,'testId_classes.npy')))
+        self.args.n_tag = len(np.load(os.path.join(self.args.asset_dir,'KnowledgeTag_classes.npy')))
+        
+        columns = ['userID', 'assessmentItemID', 'testId', 'answerCode', 'KnowledgeTag']
+        group = df[columns].groupby('userID').apply(
+                lambda r: (
+                    r['testId'].values, 
+                    r['assessmentItemID'].values,
+                    r['KnowledgeTag'].values,
+                    r['answerCode'].values
+                )
+            )
+
+        return group.values
 
     def load_data_from_file(self, file_name, is_train=True):
         csv_file_path = os.path.join(self.args.data_dir, file_name)
@@ -86,11 +106,9 @@ class Preprocess:
 
         # 추후 feature를 embedding할 시에 embedding_layer의 input 크기를 결정할때 사용
 
-                
         self.args.n_questions = len(np.load(os.path.join(self.args.asset_dir,'assessmentItemID_classes.npy')))
         self.args.n_test = len(np.load(os.path.join(self.args.asset_dir,'testId_classes.npy')))
         self.args.n_tag = len(np.load(os.path.join(self.args.asset_dir,'KnowledgeTag_classes.npy')))
-        
 
 
         df = df.sort_values(by=['userID','Timestamp'], axis=0)
@@ -106,13 +124,15 @@ class Preprocess:
 
         return group.values
 
-    def load_train_data(self, file_name):
+    def load_test_data(self, data):
+        self.test_data = self.load_data(data)
+
+    def load_train_data_csv(self, file_name):
         self.train_data = self.load_data_from_file(file_name)
 
-    def load_test_data(self, file_name):
-        self.test_data = self.load_data_from_file(file_name, is_train= False)
-
-
+    def load_test_data_csv(self, file_name):
+        self.train_data = self.load_data_from_file(file_name, is_train=False)
+        
 class DKTDataset(torch.utils.data.Dataset):
     def __init__(self, data, args):
         self.data = data
@@ -151,7 +171,7 @@ class DKTDataset(torch.utils.data.Dataset):
         return len(self.data)
 
 
-from torch.nn.utils.rnn import pad_sequence
+
 
 def collate(batch):
     col_n = len(batch[0])
